Parse quoted payment config keys
